Helper references reorg
* Formalize the reference creation pattern. `SomeReference.create` is
  responsible for determining "const-ness" and apply the right
  optimization if applicable.

* Move specialized reference out of the `/utils/references.js` file.
  This file should be reserved for common "util" references like
  `RootReference` and `CachedReference`. The specialized ones could just
  live in where they are used.

* New optimization: simple helpers with const args produces const
  references – that means things like `{{concat "foo" "bar"}}` (or more
  likely – `{{concat foo bar}}` where `foo` and `bar` are known to come
  from const sources) does not incur any updating cost.

  This implements the optimization discussed in https://github.com/tildeio/glimmer/pull/133#discussion_r60147755
